Change qty in the cart

+++ frontend/src/features/cart/cartService.js
@@ -7,9 +7,7 @@ const addItem = async (id, qty, cartItems) => {
     const response = await axios.get(`/api/products/${id}`)
     const data = response.data
 
-    /*
-    if(response.data){
-        const itemData = {
+    const item = {
             product: data._id,
             name: data.name,
             image: data.image,
@@ -17,11 +15,23 @@ const addItem = async (id, qty, cartItems) => {
             countInStock: data.countInStock,
             qty
         }
-        cartItems.push(itemData)
-        localStorage.setItem('cartItems', JSON.stringify(cartItems))
-    }*/
 
-    return itemData
+    const existItem = cartItems.find(x => x.product === item.product)
+
+    if(existItem){
+        console.log('The item already exists')
+        //Replace
+        cartItems = cartItems.map(x => x.product === existItem.product ? item : x)
+        localStorage.setItem('cartItems', JSON.stringify(cartItems))
+        return cartItems
+    } else {
+        console.log('The item doesnt exists')
+        cartItems = [...cartItems, item]
+        localStorage.setItem('cartItems', JSON.stringify(cartItems))
+        return cartItems
+    }
+
+    
 }
 
 const cartService = {
